Use Double as literal type instead of Float

+++ Sources/TEL/AST.swift
@@ -30,7 +30,7 @@ public enum Role {
 
 public enum Constant {
     case int(Int)
-    case float(Float)
+    case float(Double)
 }
 
 public indirect enum Expression {
